Revert "Merge pull request #16 from LBHackney-IT/MTTL-2794"
This reverts commit 4fdbeb13cba5d922aa4810b56711eabf288c5941, reversing
changes made to b6e013a0f6e3152bacd9c9debd29e1eb6eb353fc.
