+++ MtfhReportingDataListener.Tests/UseCase/TenureUseCaseTests.cs
@@ -15,7 +15,6 @@ using System.Linq;
 using System.Collections.Generic;
 using System.Text.Json;
 using Avro.Generic;
-using MtfhReportingDataListener.Helper;
 
 namespace MtfhReportingDataListener.Tests.UseCase
 {
@@ -25,7 +24,6 @@ namespace MtfhReportingDataListener.Tests.UseCase
         private readonly Mock<ITenureInfoApiGateway> _mockGateway;
         private readonly Mock<IKafkaGateway> _mockKafka;
         private readonly Mock<ISchemaRegistry> _mockSchemaRegistry;
-        private readonly Mock<IConvertToAvroHelper> _mockConvertToAvroHelper;
         private readonly TenureUseCase _sut;
         private readonly TenureResponseObject _tenure;
 
@@ -43,8 +41,7 @@ namespace MtfhReportingDataListener.Tests.UseCase
             _mockGateway = new Mock<ITenureInfoApiGateway>();
             _mockKafka = new Mock<IKafkaGateway>();
             _mockSchemaRegistry = new Mock<ISchemaRegistry>();
-            _mockConvertToAvroHelper = new Mock<IConvertToAvroHelper>();
-            _sut = new TenureUseCase(_mockGateway.Object, _mockKafka.Object, _mockSchemaRegistry.Object, _mockConvertToAvroHelper.Object);
+            _sut = new TenureUseCase(_mockGateway.Object, _mockKafka.Object, _mockSchemaRegistry.Object);
 
 
             _tenure = CreateTenure();
@@ -111,7 +108,7 @@ namespace MtfhReportingDataListener.Tests.UseCase
             _mockSchemaRegistry.Setup(x => x.GetSchemaForTopic(schemaName)).ReturnsAsync(mockSchemaResponse).Verifiable();
 
             await _sut.ProcessMessageAsync(_tenureCreatedMessage).ConfigureAwait(false);
-            _mockSchemaRegistry.Verify(x => x.GetSchemaForTopic(schemaName), Times.Once);
+            _mockSchemaRegistry.Verify();
         }
 
         [Fact]
@@ -189,6 +186,358 @@ namespace MtfhReportingDataListener.Tests.UseCase
 
             await _sut.ProcessMessageAsync(_tenureUpdatedMessage).ConfigureAwait(false);
             _mockKafka.Verify(x => x.SendDataToKafka(schemaName, It.IsAny<GenericRecord>()), Times.Once);
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanSetOneStringValueToAGenericRecord()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                   {
+                     ""name"": ""Id"",
+                     ""type"": ""string"",
+                     ""logicalType"": ""uuid""
+                   }
+                ]
+            }";
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, _tenure);
+
+            Assert.Equal(_tenure.Id.ToString(), receivedRecord["Id"]);
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanSetMultipleStringsToAGenericRecord()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                   {
+                     ""name"": ""Id"",
+                     ""type"": ""string"",
+                     ""logicalType"": ""uuid""
+                   },
+                   {
+                     ""name"": ""PaymentReference"",
+                     ""type"": ""string""
+                   },
+                ]
+            }";
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, _tenure);
+
+            Assert.Equal(_tenure.Id.ToString(), receivedRecord["Id"]);
+            Assert.Equal(_tenure.PaymentReference, receivedRecord["PaymentReference"]);
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanConvertDatesToUnixTimestamps()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                   {
+                     ""name"": ""Id"",
+                     ""type"": ""string"",
+                     ""logicalType"": ""uuid""
+                   },
+                   {
+                     ""name"": ""SuccessionDate"",
+                     ""type"": [""null"", ""int""]
+                   },
+                ]
+            }";
+
+            var tenure = _tenure;
+            tenure.SuccessionDate = new DateTime(1970, 01, 02);
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, tenure);
+
+            Assert.Equal(86400, receivedRecord["SuccessionDate"]);
+        }
+
+        [Theory]
+        [InlineData("IsTenanted")]
+        public void BuildTenureRecordCanSetBooleanTypeValuesToAGenericRecord(string nullableBoolFieldName)
+        {
+            var schema = @$"{{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                   {{
+                     ""name"": ""IsActive"",
+                     ""type"": ""boolean""
+                   }},
+                   {{
+                     ""name"": ""{nullableBoolFieldName}"",
+                     ""type"": [""boolean"", ""null""]
+                   }}
+                ]
+            }}";
+
+            var fieldValue = GetFieldValueFromStringName<bool>(nullableBoolFieldName, _tenure);
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, _tenure);
+
+            Assert.Equal(_tenure.IsActive, receivedRecord["IsActive"]);
+            Assert.Equal(fieldValue, receivedRecord[nullableBoolFieldName]);
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanSetNestedFields()
+        {
+            var schema = @"{
+                    ""type"": ""record"",
+                    ""name"": ""TenureInformation"",
+                    ""fields"": [
+                       {
+                         ""name"": ""TenureType"",
+                         ""type"": {
+                            ""type"": ""record"",
+                            ""name"": ""charge"",
+                            ""fields"": [
+                            {
+                                ""name"": ""Code"",
+                                ""type"": ""string""
+                            }]
+                            }
+                        }
+                    ]
+                }";
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, _tenure);
+            var receivedTenureType = (GenericRecord) receivedRecord["TenureType"];
+
+            Assert.Equal(_tenure.TenureType.Code, receivedTenureType["Code"]);
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanSetLists()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                    {
+                        ""name"": ""HouseholdMembers"",
+                        ""type"": {
+                            ""type"": ""array"",
+                            ""items"": {
+                                ""name"": ""HouseholdMember"",
+                                ""type"": ""record"",
+                                ""fields"": [
+                                    {
+                                        ""name"": ""Id"",
+                                        ""type"": ""string"",
+                                        ""logicalType"": ""uuid""
+                                    }
+                                ]
+                            }
+                        }
+                    }
+                ]
+            }";
+
+            var tenure = _tenure;
+            tenure.HouseholdMembers = new List<HouseholdMembers> { tenure.HouseholdMembers.First() };
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, tenure);
+            receivedRecord["HouseholdMembers"].Should().BeOfType<GenericRecord[]>();
+
+            var firstRecord = ((GenericRecord[]) receivedRecord["HouseholdMembers"]).ToList().First();
+            firstRecord["Id"].Should().Be(tenure.HouseholdMembers.First().Id.ToString());
+
+        }
+
+        [Fact]
+        public void BuildTenureRecordCanAssignEnums()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                    {
+                        ""name"": ""HouseholdMembers"",
+                        ""type"": {
+                            ""type"": ""array"",
+                            ""items"": {
+                                ""name"": ""HouseholdMember"",
+                                ""type"": ""record"",
+                                ""fields"": [
+                                    {
+                                        ""name"": ""Id"",
+                                        ""type"": ""string"",
+                                        ""logicalType"": ""uuid""
+                                    },
+                                    {
+                                        ""name"": ""Type"",
+                                        ""type"": {
+                                            ""name"": ""HouseholdMembersType"",
+                                            ""type"": ""enum"",
+                                            ""symbols"": [
+                                                ""Person"",
+                                                ""Organisation""
+                                            ]
+                                        }
+                                    }
+                                ]
+                            }
+                        }
+                    }
+                ]
+            }";
+
+            var tenure = _tenure;
+            tenure.HouseholdMembers = new List<HouseholdMembers> { tenure.HouseholdMembers.First() };
+            var receivedRecord = ExecuteBuildTenureRecord(schema, tenure);
+            var receivedHouseholdMember = ((GenericRecord[]) receivedRecord["HouseholdMembers"])[0];
+
+            receivedHouseholdMember["Id"].Should().Be(_tenure.HouseholdMembers.First().Id.ToString());
+            ((GenericEnum) receivedHouseholdMember["Type"]).Value.Should().Be(_tenure.HouseholdMembers.First().Type.ToString());
+        }
+
+        [Fact]
+        public void BuildTenureCanHandleNestedObjects()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                    {
+                    ""name"": ""TenuredAsset"",
+                    ""type"": {
+                        ""type"": ""record"",
+                        ""name"": ""TenuredAsset"",
+                        ""fields"": [
+                        {
+                            ""name"": ""Id"",
+                            ""type"": ""string"",
+                            ""logicalType"": ""uuid""
+                        },
+                        {
+                            ""name"": ""Type"",
+                            ""type"": [{
+                            ""name"": ""TenuredAssetType"",
+                            ""type"": ""enum"",
+                            ""symbols"": [
+                                ""Block"",
+                                ""Concierge"",
+                                ""Dwelling"",
+                                ""LettableNonDwelling"",
+                                ""MediumRiseBlock"",
+                                ""NA"",
+                                ""TravellerSite""
+                            ]
+                            }, ""null""]
+                        },
+                        ]
+                    }
+                }
+                ]
+            }";
+
+            var receivedRecord = ExecuteBuildTenureRecord(schema, _tenure);
+            var receivedRecordEnum = (GenericEnum) ((GenericRecord) receivedRecord["TenuredAsset"])["Type"];
+            receivedRecord["TenuredAsset"].Should().BeOfType<GenericRecord>();
+
+            ((GenericRecord) receivedRecord["TenuredAsset"])["Id"].Should().Be(_tenure.TenuredAsset.Id.ToString());
+            receivedRecordEnum.Value.Should().Be(_tenure.TenuredAsset.Type.ToString());
+
+        }
+
+        [Fact]
+        public void LogsOutSchemaFieldNameWhenItDoesNotExistInTenure()
+        {
+            var schema = @"{
+                ""type"": ""record"",
+                ""name"": ""TenureInformation"",
+                ""fields"": [
+                   {
+                     ""name"": ""FieldNameNotInTenure"",
+                     ""type"": ""string"",
+                   },
+                ]
+            }";
+
+            Func<GenericRecord> receivedRecord = () => ExecuteBuildTenureRecord(schema, _tenure);
+
+            receivedRecord.Should().NotThrow<NullReferenceException>();
+        }
+
+        private T GetFieldValueFromStringName<T>(string fieldName, TenureResponseObject tenure)
+        {
+            return (T) typeof(TenureResponseObject).GetProperty(fieldName).GetValue(tenure);
+        }
+
+        private GenericRecord ExecuteBuildTenureRecord(string tenureSchema, TenureResponseObject tenure)
+        {
+            var schema = @$"{{
+                ""type"": ""record"",
+                ""name"": ""TenureAPIChangeEvent"",
+                ""namespace"": ""MMH"",
+                ""fields"": [
+                    {{
+                        ""name"": ""Id"",
+                        ""type"": ""string"",
+                        ""logicalType"": ""uuid""
+                    }},
+                    {{
+                        ""name"": ""EventType"",
+                        ""type"": ""string""
+                    }},
+                    {{
+                        ""name"": ""SourceDomain"",
+                        ""type"": ""string""
+                    }},
+                    {{
+                        ""name"": ""SourceSystem"",
+                        ""type"": ""string""
+                    }},
+                    {{
+                        ""name"": ""Version"",
+                        ""type"": ""string""
+                    }},
+                    {{
+                        ""name"": ""CorrelationId"",
+                        ""type"": ""string"",
+                        ""logicalType"": ""uuid""
+                    }},
+                    {{
+                        ""name"": ""DateTime"",
+                        ""type"": ""int"",
+                        ""logicalType"": ""date""
+                    }},
+                    {{
+                        ""name"": ""User"",
+                        ""type"": {{
+                            ""type"": ""record"",
+                            ""name"": ""User"",
+                            ""fields"": [
+                                {{
+                                    ""name"": ""Name"",
+                                    ""type"": ""string""
+                                }},
+                                {{
+                                    ""name"": ""Email"",
+                                    ""type"": ""string""
+                                }}
+                            ]
+                        }}
+                    }},
+                    {{
+                        ""name"": ""Tenure"",
+                        ""type"": {tenureSchema}
+                    }}
+                ]
+            }}";
+            var tenureChangeEvent = _fixture.Create<TenureEvent>();
+            tenureChangeEvent.Tenure = _tenure;
+            var record = _sut.BuildTenureRecord(schema, tenureChangeEvent);
+            return (GenericRecord) record["Tenure"];
         }
