--- MtfhReportingDataListener/Domain/User.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace MtfhReportingDataListener.Domain
-{
-    public class User
-    {
-        public string Name { get; set; }
-
-        public string Email { get; set; }
-    }
-}
