--- MtfhReportingDataListener/UseCase/Interfaces/IContactDetailUseCase.cs
@@ -1,5 +1,0 @@
-namespace MtfhReportingDataListener.UseCase.Interfaces
-{
-    public interface IContactDetailUseCase : IMessageProcessing
-    { }
-}
