+++ MtfhReportingDataListener/UseCase/TenureUseCase.cs
@@ -13,7 +13,6 @@ using System.Linq;
 using System.Text.Json;
 using System.Reflection;
 using Newtonsoft.Json;
-using MtfhReportingDataListener.Helper;
 
 namespace MtfhReportingDataListener.UseCase
 {
@@ -22,15 +21,12 @@ namespace MtfhReportingDataListener.UseCase
         private readonly ITenureInfoApiGateway _tenureInfoApi;
         private readonly IKafkaGateway _kafkaGateway;
         private readonly ISchemaRegistry _schemaRegistry;
-        private readonly IConvertToAvroHelper _convertToAvroHelper;
 
-        public TenureUseCase(ITenureInfoApiGateway gateway, IKafkaGateway kafkaGateway,
-                             ISchemaRegistry schemaRegistry, IConvertToAvroHelper convertToAvroHelper)
+        public TenureUseCase(ITenureInfoApiGateway gateway, IKafkaGateway kafkaGateway, ISchemaRegistry schemaRegistry)
         {
             _tenureInfoApi = gateway;
             _kafkaGateway = kafkaGateway;
             _schemaRegistry = schemaRegistry;
-            _convertToAvroHelper = convertToAvroHelper;
         }
 
         [LogCall]
@@ -61,10 +57,98 @@ namespace MtfhReportingDataListener.UseCase
 
             var topicName = Environment.GetEnvironmentVariable("TENURE_SCHEMA_NAME");
             var schema = await _schemaRegistry.GetSchemaForTopic(topicName);
-            var record = _convertToAvroHelper.BuildRecord(schema, tenureChangeEvent);
+            var record = BuildTenureRecord(schema, tenureChangeEvent);
 
             await _kafkaGateway.CreateKafkaTopic(topicName);
             _kafkaGateway.SendDataToKafka(topicName, record);
+        }
+
+        public GenericRecord BuildTenureRecord(string schema, TenureEvent tenureResponse)
+        {
+            return PopulateFields(tenureResponse, Schema.Parse(schema));
+        }
+
+        public GenericRecord PopulateFields(object item, Schema schema)
+        {
+            var record = new GenericRecord((RecordSchema) schema);
+            ((RecordSchema) schema).Fields.ForEach(field =>
+            {
+                PropertyInfo propInfo = item.GetType().GetProperty(field.Name);
+                if (propInfo == null)
+                {
+                    Console.WriteLine($"Field name: {field.Name} not found in {item} with type {item.GetType()}");
+                    return;
+                }
+
+                var fieldValue = propInfo.GetValue(item);
+                var fieldSchema = field.Schema;
+                var fieldType = field.Schema.Tag;
+
+                if (fieldType == Schema.Type.Union)
+                {
+                    if (fieldValue == null)
+                    {
+                        record.Add(field.Name, null);
+                        return;
+                    }
+                    fieldSchema = GetNonNullablePartOfNullableSchema(field.Schema);
+                    fieldType = fieldSchema.Tag;
+                }
+
+                if (fieldType == Schema.Type.String)
+                {
+                    record.Add(field.Name, fieldValue.ToString());
+                }
+                else if (fieldType == Schema.Type.Enumeration)
+                {
+                    record.Add(field.Name, new GenericEnum((EnumSchema) fieldSchema, fieldValue.ToString()));
+                }
+                else if (fieldValue.GetType() == typeof(DateTime))
+                {
+                    record.Add(field.Name, UnixTimestampNullable(fieldValue));
+                }
+                else if (fieldType == Schema.Type.Array)
+                {
+                    var itemsType = fieldValue.GetType().GetGenericArguments().FirstOrDefault();
+                    var itemsSchema = GetSchemaForArrayItems(fieldSchema);
+
+                    var fieldValueAsList = ((JsonElement) System.Text.Json.JsonSerializer.Deserialize<object>(JsonConvert.SerializeObject(fieldValue))).EnumerateArray();
+                    var recordsList = fieldValueAsList.Select(listItem => PopulateFields(JsonConvert.DeserializeObject(listItem.ToString(), itemsType), itemsSchema)).ToArray();
+
+                    record.Add(field.Name, recordsList);
+                }
+                else if (fieldType == Schema.Type.Record)
+                {
+                    record.Add(field.Name, PopulateFields(fieldValue, fieldSchema));
+                }
+                else
+                {
+                    record.Add(field.Name, fieldValue);
+                }
+            });
+
+            return record;
+        }
+
+        private Schema GetSchemaForArrayItems(Schema arraySchema)
+        {
+            var jsonSchema = (JsonElement) System.Text.Json.JsonSerializer.Deserialize<object>(arraySchema.ToString());
+            jsonSchema.TryGetProperty("items", out var itemsSchemaJson);
+            return Schema.Parse(itemsSchemaJson.ToString());
+        }
+
+        private Schema GetNonNullablePartOfNullableSchema(Schema nullableSchema)
+        {
+            var jsonSchema = (JsonElement) System.Text.Json.JsonSerializer.Deserialize<object>(nullableSchema.ToString());
+            jsonSchema.TryGetProperty("type", out var unionList);
+            var notNullSchema = unionList.EnumerateArray().First(type => type.ToString() != "null").ToString();
+            return Schema.Parse(notNullSchema);
+        }
+
+        private int? UnixTimestampNullable(object obj)
+        {
+            var date = (DateTime?) obj;
+            return (int?) (date?.Subtract(new DateTime(1970, 1, 1)))?.TotalSeconds;
         }
     }
 }
